# Conflicts:
#	examples/animation/assimp_bones.cpp
#	examples/controls/transform.cpp
#	examples/demo.cpp
#	examples/geometries/box_geometry.cpp
#	examples/geometries/cylinder_geometry.cpp
#	examples/geometries/heightmap.cpp
#	examples/geometries/plane_geometry.cpp
#	examples/geometries/sphere_geometry.cpp
#	examples/helpers/depth_sensor.cpp
#	examples/misc/CMakeLists.txt
#	examples/objects/instancing.cpp
#	examples/objects/lod.cpp
#	examples/objects/sprite.cpp
#	examples/objects/text_sprite.cpp
#	examples/projects/Crane3R/main.cpp
#	examples/projects/InvertedPendulum/main.cpp
#	examples/projects/MotorControl/main.cpp
#	examples/projects/Snake/main.cpp
#	examples/projects/Youbot/youbot.cpp
#	examples/projects/Youbot/youbot_kine.cpp
#	examples/textures/cubemap.cpp
#	examples/textures/texture3d.cpp
#	include/threepp/helpers/DepthSensor.hpp
#	include/threepp/objects/HUD.hpp
#	src/threepp/canvas/Canvas.cpp
#	src/threepp/objects/HUD.cpp

+++ src/threepp/animation/KeyframeTrack.cpp
@@ -59,7 +59,7 @@ KeyframeTrack& KeyframeTrack::trim(float startTime, float endTime) {
     int from = 0;
     int to = static_cast<int>(nKeys) - 1;
 
-    while (from != nKeys && times_[from] < startTime) {
+    while (static_cast<size_t>(from) != nKeys && times_[from] < startTime) {
 
         ++from;
     }
@@ -71,7 +71,7 @@ KeyframeTrack& KeyframeTrack::trim(float startTime, float endTime) {
 
     ++to;// inclusive -> exclusive bound
 
-    if (from != 0 || to != nKeys) {
+    if (from != 0 || static_cast<size_t>(to) != nKeys) {
 
         // empty tracks are forbidden, so keep at least one keyframe
         if (from >= to) {
@@ -173,7 +173,7 @@ KeyframeTrack& KeyframeTrack::optimize() {
         ++writeIndex;
     }
 
-    if (writeIndex != times.size()) {
+    if (static_cast<size_t>(writeIndex) != times.size()) {
 
         this->times_ = AnimationUtils::arraySlice(times, 0, writeIndex);
         this->values_ = AnimationUtils::arraySlice(values, 0, writeIndex * stride);
